Release 3.0
+ Change of the organization
+ Added the algorithm to draw figure in a grid with one input and one output by recursion
+ Added the algorithm to draw a figure in a grid with numbers
- Delete "idee.md" file
