+++ Draw_random_grid/console.py
