--- Idée/Algo_retirer_chiffre.md
@@ -1,6 +1,0 @@
-x Y x       //90° 
-
-x / x
-/ Y /       //45°
-x / x
-
